+++ src/index.ts
@@ -5,6 +5,12 @@ export { WakuMessage } from './lib/waku_message';
 
 export { ChatMessage } from './lib/chat_message';
 
+export {
+  WakuLightPush,
+  LightPushCodec,
+  PushResponse,
+} from './lib/waku_light_push';
+
 export { WakuRelay, RelayCodec } from './lib/waku_relay';
 
 export { WakuStore, StoreCodec } from './lib/waku_store';